Move LicenseNotification into own module
